requirements_python.txt
matplotlib==3.7.0
numpy==1.26.2
pandas==1.5.3
Pillow==10.1.0
powerlaw==1.5
pytesseract==0.3.10
pytz==2022.7.1
scikit_learn==1.2.1
scipy==1.11.4
seaborn==0.13.0
squarify==0.4.3
statsmodels==0.13.5
tqdm==4.64.1